update auth delete sauce

+++ backend/controllers/sauce.js
@@ -95,14 +95,20 @@ exports.modifySauce = (req, res, next) => {
         .catch((error) => res.status(500).json({ error }));
 };
 exports.deleteSauce = (req, res, next) => {
-    Sauce.findOne({ _id: req.params.id })
-        .then((sauce) => {
-            const filename = sauce.imageUrl.split("/images/")[1];
-            fs.unlink(`images/${filename}`, () => {
-                Sauce.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: "OK" }))
-                    .catch((error) => res.status(400).json({ error }));
-            });
-        })
-        .catch((error) => res.status(500).json({ error }));
+    Sauce.findOne({ _id: req.params.id }).then((sauce) => {
+        if (!sauce) {
+            return res.status(404).json({ error: "Sauce introuvable" });
+        }
+        if (sauce.userId !== req.auth.userId) {
+            return res
+                .status(401)
+                .json({ error: "Vous n'avez pas les droits" });
+        }
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+            Sauce.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({ message: "Deleted" }))
+                .catch((error) => res.status(400).json({ error }));
+        });
+    });
 };
